Modify script.js
I added a new class, some new variables and a new eventListener. Modified
two functions:

* New Class
Added the class "firstTimeClicked". This class gets added in the DOM only
if one of the operators (+, -, *, /) is clicked before any of the numbers
being clicked.

* New Variables
Added variables "displayValueOne", "displayValueTwo" and "operator".
These variables are used to store the values and the operator that the
user enters. After storing them, these values are entered as arguments
inside the operate() function, in order to calculate the result.

* New eventListener
Added an eventListener on the Enter button. When Enter gets clicked, a
calculation is performed and the result is shown on display.

* Modified 2 Functions
Modified function populateDisplay() and function showOperator().